[style] Position 테스트 파일 분리
